Creating the getDataElement() function
This function will be used in all pages to retrieves elements that will
be replaced using JavaScript.

It makes sure the element is retrieved and throws an error otherwise.
This function avoids repeating the existence assertion when querying a
element (DRY).

It also marks the importance of using the `data` attribute on elements that will be edited by JS files.
Thus, when editing the HTML file, one must be careful with element bearing a
`data` attribute.